feat(fantoken): remove burnedCoins
