+++ proto/bitsong/fantoken/v1beta1/query.proto
@@ -12,21 +12,20 @@ option go_package = "github.com/bitsongofficial/go-bitsong/x/fantoken/types";
 
 // Query creates service with fantoken as RPC
 service Query {
+
   // FanToken returns fantoken with fantoken name
   rpc FanToken(QueryFanTokenRequest) returns (QueryFanTokenResponse) {
     option (google.api.http).get = "/bitsong/fantoken/v1beta1/denom/{denom}";
   }
+
   // FanTokens returns the fantoken list
   rpc FanTokens(QueryFanTokensRequest) returns (QueryFanTokensResponse) {
     option (google.api.http).get = "/bitsong/fantoken/v1beta1/fantokens";
   }
+
   // Params queries the fantoken parameters
   rpc Params(QueryParamsRequest) returns (QueryParamsResponse) {
     option (google.api.http).get = "/bitsong/fantoken/v1beta1/params";
-  }
-  // TotalBurn queries all the burnt coins
-  rpc TotalBurn(QueryTotalBurnRequest) returns (QueryTotalBurnResponse) {
-    option (google.api.http).get = "/bitsong/fantoken/v1beta1/total_burn";
   }
 }
 
@@ -55,13 +54,4 @@ message QueryParamsRequest {}
 // QueryParametersResponse is response type for the Query/Parameters RPC method
 message QueryParamsResponse {
   bitsong.fantoken.v1beta1.Params params = 1 [ (gogoproto.nullable) = false ];
-}
-
-// QueryTotalBurnRequest is request type for the Query/TotalBurn RPC method
-message QueryTotalBurnRequest {}
-
-// QueryTotalBurnResponse is response type for the Query/TotalBurn RPC method
-message QueryTotalBurnResponse {
-  repeated cosmos.base.v1beta1.Coin burned_coins = 1
-      [ (gogoproto.nullable) = false ];
 }
