+++ x/fantoken/keeper/grpc_query.go
@@ -107,10 +107,3 @@ func (k Keeper) Params(c context.Context, req *types.QueryParamsRequest) (*types
 
 	return &types.QueryParamsResponse{Params: params}, nil
 }
-
-// TotalBurn return the all Burn coin
-func (k Keeper) TotalBurn(c context.Context, req *types.QueryTotalBurnRequest) (*types.QueryTotalBurnResponse, error) {
-	ctx := sdk.UnwrapSDKContext(c)
-
-	return &types.QueryTotalBurnResponse{BurnedCoins: k.GetAllBurnedCoins(ctx)}, nil
-}
