+++ x/fantoken/keeper/store.go
@@ -14,45 +14,6 @@ func (k Keeper) HasFanToken(ctx sdk.Context, denom string) bool {
 	return store.Has(types.KeyDenom(denom))
 }
 
-func (k Keeper) SetBurnCoin(ctx sdk.Context, total sdk.Coin) {
-	bz := k.cdc.MustMarshal(&total)
-	key := types.KeyBurnFanTokenAmt(total.Denom)
-
-	store := ctx.KVStore(k.storeKey)
-	store.Set(key, bz)
-}
-
-// GetBurnedCoins returns the total amount of the burned fantoken
-func (k Keeper) GetBurnedCoins(ctx sdk.Context, denom string) sdk.Coin {
-	key := types.KeyBurnFanTokenAmt(denom)
-	store := ctx.KVStore(k.storeKey)
-	bz := store.Get(key)
-
-	if len(bz) == 0 {
-		return sdk.Coin{Denom: denom, Amount: sdk.ZeroInt()}
-	}
-
-	var coin sdk.Coin
-	k.cdc.MustUnmarshal(bz, &coin)
-
-	return coin
-}
-
-// GetAllBurnedCoins returns the total amount of all the burned fantokens
-func (k Keeper) GetAllBurnedCoins(ctx sdk.Context) []sdk.Coin {
-	store := ctx.KVStore(k.storeKey)
-
-	var coins []sdk.Coin
-	it := sdk.KVStorePrefixIterator(store, types.PefixBurnFanTokenAmt)
-	for ; it.Valid(); it.Next() {
-		var coin sdk.Coin
-		k.cdc.MustUnmarshal(it.Value(), &coin)
-		coins = append(coins, coin)
-	}
-
-	return coins
-}
-
 func (k Keeper) setWithMetadataAuthority(ctx sdk.Context, owner sdk.AccAddress, denom string) {
 	store := ctx.KVStore(k.storeKey)
 	bz := k.cdc.MustMarshal(&gogotypes.StringValue{Value: denom})
