+++ x/fantoken/types/genesis_test.go
@@ -36,7 +36,6 @@ func TestValidateGenesis(t *testing.T) {
 						},
 					},
 				},
-				BurnedCoins: nil,
 			},
 			valid: true,
 		},
@@ -59,7 +58,6 @@ func TestValidateGenesis(t *testing.T) {
 						},
 					},
 				},
-				BurnedCoins: nil,
 			},
 			valid: true,
 		},
@@ -82,7 +80,6 @@ func TestValidateGenesis(t *testing.T) {
 						},
 					},
 				},
-				BurnedCoins: nil,
 			},
 			valid: true,
 		},
@@ -94,8 +91,7 @@ func TestValidateGenesis(t *testing.T) {
 					MintFee:  sdk.NewCoin(sdk.DefaultBondDenom, sdk.ZeroInt()),
 					BurnFee:  sdk.NewCoin(sdk.DefaultBondDenom, sdk.ZeroInt()),
 				},
-				FanTokens:   nil,
-				BurnedCoins: nil,
+				FanTokens: nil,
 			},
 			valid: true,
 		},
@@ -114,7 +110,6 @@ func TestValidateGenesis(t *testing.T) {
 						MetaData:  Metadata{},
 					},
 				},
-				BurnedCoins: nil,
 			},
 			valid: false,
 		},
@@ -136,7 +131,6 @@ func TestValidateGenesis(t *testing.T) {
 						},
 					},
 				},
-				BurnedCoins: nil,
 			},
 			valid: false,
 		},
@@ -158,7 +152,6 @@ func TestValidateGenesis(t *testing.T) {
 						},
 					},
 				},
-				BurnedCoins: nil,
 			},
 			valid: true,
 		},
@@ -181,7 +174,6 @@ func TestValidateGenesis(t *testing.T) {
 						},
 					},
 				},
-				BurnedCoins: nil,
 			},
 			valid: true,
 		},
