+++ x/fantoken/types/keys.go
@@ -21,9 +21,6 @@ var (
 
 	// PrefixFanTokens defines a prefix for the fan tokens
 	PrefixFanTokens = []byte{0x02}
-
-	// PeffxBurnFanTokenAmt defines a prefix for the amount of fan token burnt
-	PefixBurnFanTokenAmt = []byte{0x03}
 )
 
 // KeyDenom returns the key of the token with the specified denom
@@ -34,9 +31,4 @@ func KeyDenom(denom string) []byte {
 // KeyFanTokens returns the key of the specified owner and denom. Intended for querying all fan tokens of an owner
 func KeyFanTokens(owner sdk.AccAddress, denom string) []byte {
 	return append(append(PrefixFanTokens, owner.Bytes()...), []byte(denom)...)
-}
-
-// KeyBurnTokenAmt returns the key of the specified min unit.
-func KeyBurnFanTokenAmt(denom string) []byte {
-	return append(PefixBurnFanTokenAmt, []byte(denom)...)
 }
